\name{AirQualityBogota}
\alias{AirQualityBogota}
\title{Air quality data of Bogota, Colombia}
\description{
  Particular material 10 (PM10) of Bogota measure in 10 stations in the city.
}
\usage{data(AirQualityBogota)}
\format{two data.frame of measurements 'PM10' and coordinates 'coord'. Also a map file of class SpatialPolygonDataFrame load from a shape file by \code{rgdal::readOGR}}
\source{IDEAM}
\references{
IDEAM
}
\keyword{datasets}
