\name{COKMexico}
\alias{COKMexico}
\title{Air quality data of Mexico}
\description{
  Particular material 10 (PM10) and NO2 measured in 13 and 18 locations in Mexico.
}
\usage{data(COKMexico)}
\format{four data.frame of measurements 'PM10','NO2' and their coordinates 'coord_PM10' and
'coord_NO2'. Also a map file of class SpatialPolygonDataFrame load from a shape file by \code{rgdal::readOGR}}
\source{Mexico}
\references{
Mexico
}
\keyword{datasets}
