\name{coord}
\docType{data}
\alias{coord}
\title{Coordinates of measurement stations Bogota, Colombia}
\description{
  Coordinates of 10 stations in the city.
}
\usage{data(AirQualityBogota)}
\format{data.frame of coordinates}
\source{Secretaria de Ambiente de Bogotá}
\references{
  Monitor network of air quality of Bogota \code{http://rmcab.ambientebogota.gov.co}
}
\keyword{datasets}
