\name{map_mex}
\docType{data}
\alias{map_mex}
\title{map of Mexico}
\description{
  Map of Mexico.
}
\usage{data(COKMexico)}
\format{Map file of class SpatialPolygonDataFrame}
\source{Automatic Monitoring System SEDEMA}
\references{
  \code{http://www.aire.df.gob.mx/default.php}
}
\keyword{datasets}