\name{PM10}
\docType{data}
\alias{PM10}
\title{PM10 of Bogota, Colombia}
\description{
  Particular material 10 (PM10) of Bogota measure in 10 stations in the city.
}
\usage{data(AirQualityBogota)}
\format{data.frame of measurements 'PM10'}
\source{Secretaria de Ambiente de Bogotá}
\references{
  Monitor network of air quality of Bogota \code{http://rmcab.ambientebogota.gov.co}
}
\keyword{datasets}
