\name{SpatFD-package}
\alias{SpatFD-package}
\alias{SpatFD}
\docType{package}
\title{
\packageTitle{SpatFD}
}
\description{
\packageDescription{SpatFD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpatFD}
\packageIndices{SpatFD}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SpatFD}

Maintainer: \packageMaintainer{SpatFD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
