\name{vowels}
\alias{vowels}
\title{Imaginary thinking of the five Spanish vowels}
\description{
  Data consist of EEG signals taken from 21 electrodes from imaginary thinking of the five Spanish vowels,  to be applied into a BCI for a hand prosthesis.
}
\usage{data(vowels)}
\format{two data.frame of measurements 'vowels' and coordinates 'vowels_coords'.}
\source{https://github.com/carlos-sarmientov/DATABASE-IMAGINED-VOWELS-1}
\references{
  Classification techniques for imaginary speech brain signal through spatial functional data \code{https://repositorio.unal.edu.co/handle/unal/85267}
}
\keyword{datasets}
