\name{vowels_coords}
\docType{data}
\alias{vowels_coords}
\title{Coordinates of electrodes from the vowels data set}
\description{
  Coordinates of the electrodes from the vowels data set.
}
\usage{data(vowels)}
\format{matrix of coordinates}
\source{https://github.com/carlos-sarmientov/DATABASE-IMAGINED-VOWELS-1}
\references{
  Classification techniques for imaginary speech brain signal through spatial functional data \code{https://repositorio.unal.edu.co/handle/unal/85267}
}
\keyword{datasets}
