\name{COKMexico}
\docType{data}
\alias{COKMexico}
\title{Air quality data of Mexico}
\description{
  Particular material 10 (PM10) and NO2 measured in 13 and 18 locations in Mexico.
  The data correspond to consecutive hours from January 01, 2015 at 1:00 a.m. to May 30, 2015 at 12:00 a.m., at 23 environmental stations. The stations in the air quality network RAMA (Red Automática de Monitoreo Atmosférico), monitor hourly particulate matter up to 10 micrometers in size (PM10) and  Nitrogen dioxide (NO2) among others.  The particulate matter (PM) is an important component of air pollution. NO2 is a gaseous air pollutant produced by the road traffic and other fossil fuel combustion processes and it contributes to the formation and modification of other air pollutants such as particulate matter.
}
\usage{data(COKMexico)}
\format{four data.frame of measurements 'Mex_PM10','NO2' and their coordinates 'coord_PM10' and
'coord_NO2'. Also a map (map_mex) file of class SpatialPolygonDataFrame}
\source{Mexico}
\keyword{datasets}
