\name{map}
\alias{map}
\title{map of Bogota, Colombia}
\description{
  Map of Bogota.
}
\usage{data(AirQualityBogota)}
\format{Map file of class SpatialPolygonDataFrame}
\source{Unidad Administrativa Especial de Catastro Distrital}
\references{
  \code{https://datosabiertos.bogota.gov.co/dataset/sector-catastral}
}
\keyword{datasets}
